requirements.txt
snowflake-connector-python >= 3.2.0
snowflake-snowpark-python >= 1.8.0
